backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
ultralytics==8.1.47
supervision==0.19.0
scikit-learn==1.4.1.post1
numpy==1.26.4
ffmpeg-python==0.2.0
reportlab==4.1.0
Pillow==10.3.0
aiofiles==23.2.1
